Auto-generate DSL from Ecore
An Xtext DSL is created by using the Eclipse wizard based on the existing
Ecore meta-model. No further changes are made (yet). The DSL contains
elements from different and nested EPackages, so some changes will be
necessary to make this solution executable (-> further commits).
The recently created EClass CompleteArchitecture is chosen as entry rule
so that all aspects of a component-based system can be described with one
grammar and in one file.