Add CompositeComponent support to grammar
Xtext seems not to like inheritance. In the generated grammar, many of the
abstract types where just flattened (attributes added to all sub-classes).
The EClass "Component" is not abstract, but has a sub-class
CompositeComponent. A grammar rule was only generated for Component. As a
fix, this commit introduces a new rule "CompositeComponent" and renames
"Component" to "BasicComponent" (just the rule, not the Ecore element).
The rule "Component" is re-introduced, but just delegates to either
BasicComponent or CompositeComponent.