Add transformation rules
A library with transformation rules is created (which covers all EClasses
of CBSM), but no transformation script. Not tested yet. One file of the
.settings has also be to checked in, because it defines how the namespace
URIs are resolved to concrete ecore files for the CBSM. This is not
necessary for Palladio which resolves the namespaces via Palladio Eclipse
plug-ins (and adding namespace mappings to our pcm.ecore in the project
results in strange exceptions when actually creating transformations).

Assumptions (design decisions):
- Interface, Signature, ProvidedRole, RequiredRole are represented as
OperationInterface, OperationSignature, OperationProvidedRole,
OperationRequiredRole (the directly corresponding Palladio EClasses are
abstract and have multiple sub-classes).
- ComplexDataType is represented as CompositeDataType (there are two kinds
of complex types in Palladio, Composite- and CollectionDataType).
- CBSM's SimpleTypes Map, List, Date are represented as
PrimitiveDataType#String, because they have no equivalent in Palladio.
- BranchPath is represented as ProbabilisticBranchTransition (the abstract
class AbstractBranchTransition of Palladio has also a sub-class
GuardedBranchTransition, but we have to return one non-abstract instance).

ToDos:
- Palladio supports multiple SEFFs per Component, we support only one
BehaviorDescription.
- Palladio's LoopAction references a ResourceDemandingBehavior (which
contains AbstractActions), but our Loop references a concrete Action. The
meta-model should be changed here.
- Palladio's AbstractBranchTransition references a
ResourceDemandingBehavior (which contains AbstractActions), but our
BranchPath references a concrete Action. The meta-model should be changed
here.