Add two Assembly2System transformations
A simple assembly transformation gets an .assembly file and transforms the
top-level element "System" to a PCM "System" in a .system file. However,
the assembly view type is highly dependent on the repository view type,
e.g. it uses Roles and Components. In the simple transformation, those
elements cannot be resolved an remain null. Therefore, a second
transformation is added which performs a simultaneous transformation of
assembly and  repository, reusing the corresponding transformation scripts.