Add two Allocation2Allocation transformations
A simple allocation transformation gets an .allocation file and transforms
the top-level element "Allocation" to a PCM "Allocation". However,
the allocation view type is highly dependent on the assembly and
environment view type as it uses Containers and AssemblyContexts. In the
simple transformation, those elements cannot be resolved an remain null.
Therefore, a second transformation is added which performs a simultaneous
transformation of allocation, assembly and environment, reusing the
corresponding transformation scripts (and for the assembly also the
repository is used).