Make CBSM more compliant to Palladio
- ViewType extends NamedElement now, so Repository, System, Environment and
Allocation get a name as in Palladio (before this commit, the
transformation only returned a default name).
- Components can have multiple BehaviorDescriptions (before: just one), as
PCM Components can have multiple SEFFs. BehaviorDescription is also made
a NamedElement in our meta-model (but SEFF in PCM is not named).
- Action is made abstract. Non-abstract Actions are InternalAction, Loop,
ExternalCall and Branch. The task for assignment 1 only demands those four
types, so a non-abstract class Action simply is not necessary (but as
abstract super-class it serves a purpose, as it contains the EReferences
to predecessor and successor).
- Loop and BranchPath reference exactly one BehaviorDescription instead of
0..* Actions. This is similar to Palladio's AbstractBranchTransition and
AbstractLoopAction which both reference one ResourceDemandingBehavior.

This commit contains many files, as
- the meta-model is changed as described above
- the meta-model code is re-generated
- the Xtext DSL is adapted accordingly
- the Xtext files are re-generated
- the QVT-O transformation library file is adapted accordingly
- the MediaStore example (Ecore, XText, XText serialized) is adapted
accordingly

Hopefully there won't be many meta-model changes in the future ...