Fix SimpleType mapping
CBSM's SimpleType enum (SimpleTypeInstance) contains more values than the
corresponding enum of PCM's PrimitiveType. First, there is Float, but that
can be easily mapped to PCM's DOUBLE. More difficult are Date, Map and List
which really have no similar PrimitiveType in Palladio. This commit changes
the mapping of those types; instead of returning PrimitiveTypeEnum::STRING
(which was used before as a quick workaround), a CompositeDataType is
created which has the name of the SimpleTypeInstance. For Map and List,
PCM's CollectionDataType would be more intuitive as mapping, but that EClass
has a required feature "innerType" which we cannot provide, so the instance
would be invalid.