Fix some PCM-OCL-related errors
Some OCL constraints of PCM were violated when validating the transformed
files. Most of these errors are fixed now in all relevant places (CBSM meta-
model, Xtext grammar, transformation library, models):

- CBSM's CompositeComponent can contain Connectors now (exact solution:
EReference to Connectors [which only was part of the System EClass before]
is renamed and moved to the super-class CompositeProvidingAndRequiringElement.
As a consequence, the CompositeComponent in the model contains it inner
AssemblyConnectors as well as its DelegationConnectors from inner to outer
Roles (before this commit: all Connectors centrally stored in System).
- CBSM's "outerAssemblyContext" EReference of DelegationConnectors is
removed, so similar to PCM only the inner AssmemblyContext and the inner
and outer Role are referenced. This enables DelegationConnectors for the
system as a whole which don't have an outer AssemblyContext (in our
MediaStoreExample: the HTTP Interface of WebGUI is also provided by the
System, so a new ProvidedDelegationConnector is added).
- CBSM's Allocation gets EReferences to the allocated System and the
allocation Environment. OCL constraints guarantee that the AssemblyContexts
of the AllocationContexts of the Allocation belong to those System and that
the Containers of the AllocationContexts of the Allocation belong to those
Environment.

The PCM files generated by the transformation only contain one validation
error after these changes: The required feature
"communicationLinkResourceSpecifications_LinkingResource" of the EClass
LinkingResource is missing, but is is not easy to fix, as there is no
comparable class in our meta-model and the expected EClass
CommunicationLinkResourceSpecification has some references to other EClasses
which are also not in our meta-model. Therefore, we might just accept that
one validation error.