Fix SEFF mapping
The old CBSM meta-model had only one EClass "BehaviorDescription" to
represent PCM's "ResourceDemandingBehavior", always transformed to PCM's
"ResourceDemandingSEFF". However, two EReferences of a SEFF
where missing (to the described service [Signature] and the containing
component), which resulted in violated OCL constraints. These EReferences
are only interesting for SEFFs, but not the general PCM super-class
"ResourceDemandingBehaviour" which can be instantiated and also referenced
from "LoopAction" and "AbstractBranchTransition" in Palladio. As a
consequence, the CBSM meta-model gets a new EClass
"ComponentBehaviorDescription" which extends "BehaviorDescription" and has
the necessary references for representing a SEFF. The transformation rules
and the XText grammar are adapted accordingly (the MediaStore example does
not need an update because it doesn't contain any Behavior/SEFF).