Fix Link mapping
The mapping of Link to LinkingResource did not set the required feature
"communicationLinkResourceSpecifications_LinkingResource" which resulted
in a validation error. This problem has been known for a while (see
commit 659434a160), but fixing it seemed difficult because
1) no equivalent element for "CommunicationLinkResourceSpecification"
exists in the CBSM meta-model,
2) the required type itself has a lot of required features, some of them
also not part of CBSM,
3) the feature "communicationLinkResourceType" of the required resource
specification is part of (contained in) a repository in the EPackage
"resourcetype", which again has no equivalent in CBSM.

Fortunately, the Palladio plug-ins provided some default resource types
in the file "pathmap://PCM_MODELS/Palladio.resourcetype", one of them
being a "CommunicationLinkResourceType" (called "LAN"). This commit
changes the Environment2ResourceEnvironment transformation such that it
tries to access this repository, retrieving the "LAN" and passing it as
parameter to the Environment as well as Link mapping. Transformations
using the Environment2ResourceEnvironment transformation are adapted
accordingly (allocation & full transformation), the same applies to the
launch configurations. Most of the transformations' launch configurations
are also renamed to consistently contain "MediaStore" in their name.

The commit seems to contain a lot of changes in the transformed files, but
this is a result of Palladio's identifier generation (and many features
being not ordered, so the items of collections might be arranged
differently).