Create code generation project
This is the initial commit for task 4: code generation. A new project is
created and all necessary configuration steps are made (mainly Xtend and
Xtext dependencies). A very small Xtend-based generation logic is added
which creates empty Java interfaces for all CBSM Interfaces in .repository
files. The generation can be started by right-clicking on .repository files;
a new item (with a special icon) will be shown in the menu (so we use
Eclipse's extension point mechanism). This command is handled by a Java
class that mainly resolves the file path causing the event and then
delegates to the CbsmRepositoryCodeGenerator. Some dependency injection
via Guice is also done, but that's rather a technical detail.

A first approach based on a MWE2 workflow did not work as intended
(https://christiandietrich.wordpress.com/2011/07/29/xtend2-code-generators-with-non-xtext-models/)
and had some strange exceptions, therefore we chose the current solution.
Helpful resources are:
- https://christiandietrich.wordpress.com/2011/10/15/xtext-calling-the-generator-from-a-context-menu/
- https://www.eclipse.org/forums/index.php/t/628292/