Prepare example project for generated code
With the old structure of the MediaStore project, generated Java files
would only be considered as rather normal text files with some syntax
highlighting. This commit converts the project to a Java project and adds
the output folder of code generation to the class path, so all Java files
are treated "normally" (automatic compilation, error checking etc.).