Generate code for Interfaces
The code generation logic for Interfaces (and therefore Signatures,
Parameters, DataTypes) is implemented (task 4a). Each CBSM Interface
is translated into a Java interface residing in a package with the
name of the CBSM Repository. Only SimpleTypes are converted to their
direct counterpart in Java, ComplexTypes (which really could be anything)
become Objects. The handler of the code generation actions ensures that
the latest version of the .repository file is used (not a cached one) and
refreshes the project afterwards, so the user does not have to press "F5".