Generate code for components
The code generation logic for Components (and therefore ProvidedRoles,
RequiredRoles, Repository) is implemented (task 4b and c). Each component
gets as extra package, has the required roles as fields (inclusive setter
methods) and the provided roles as super-interfaces whose methods are
implemented. A helper class is generated for the whole Repository that can
perform null checks for the methods of the Components. Additional to the
task, method with return types get a statement which returns a default
value, so there are no compile errors in that case.