Fix AssemblyContext mapping
When thinking about task 5's Interfaces that extends Interfaces (which is
not part of the meta-model or any other project yet), the question arose
how such references would be resolved when writing a transformation for
them. There already is a kind of indirectly self-referencing mapping in
the QVT-O transformation: If a CompositeComponent is transformed, the
mapping for the AssemblyContexts of the CompositeComponent is called; these
represent the contained Components which are tried to be resolved. It might
happen that the CompositeComponent is placed above its contained Components
in the Repository model. In that case, the resolution attempt fails and the
attribute "encapsulatedComponent" of the AssemblyContext remains empty.
This commit fixes this by moving the final resolving of the references to
encapsulated components to the end of the transformation (late resolve), so
it is guaranteed that all Components are already transformed and can be
resolved.