Include all Repository elements in diagram
All relevant EClasses and EReferences of the Repository are included in
the diagram (meaning Components, CompositeComponents, Interfaces,
SimpleTypes, ComplexTypes, Signatures, Parameters, Roles and their
references). This commit does not adress the styling (always defaults used),
so the visual result is horrible, but complete. The operations for editing
the diagram are also not implemented yet.