Remove nodes and edges for DataTypes
The task only demands that DataTypes can be used for return types and
parameters, but they should be created in the usual tree editor. Having
separate nodes for DataTypes makes the diagram quite confusing, because
each parameter has an edge to a DataType. Therefore some alternative
representation should be found in future commits.