Add node/container creation tools
New tools are added that allow the creation of Interfaces, Components,
CompositeComponents, ProvidedRoles, RequiredRoles, Signatures and
Parameters (so all nodes and containers) which get a default name (based on
the count of existing elements of that type in the current container). The
feature can be used to drag&drop new elements directly from the palette
view to the diagram in Eclipse.