Enable type selection and improve editing
Two tools are added which allow the selection of a type for a Parameter or
a return type for a Signature (types are taken from the Repository). The
return type of Signatures and the type of a Parameter are also displayed
in the label. The label editing functionality is changed such that the user
can only edit the name in the diagram, but the "generated" information
in the label as Signature return type, kind of Component, Parameter type
are not affectected by the editing. Finally, as the SelectionWizard does
not allow no selection, a tool is added which lets the user remove the
return type (change to void).