Add icons for Repository elements
The elements of repositories which are part of the graphical editor get an
icon. The icons are inserted into the meta-model edit project (replace the
default ones) and will be automatically used in the graphical editor (for
diagram elements and the add actions in the palette) as well as the usual
tree editor. An icon of Sirius (red x) is chosen for the "remove return
type" action. New icons are created for the edge creation actions and
stored in the repositoryeditor project as well as explicitely referenced
in the *.odesign file.