Change references in *.aird file
The *.aird file in the MediaStore example project stores the different
representations, especially the diagram for the repository which was created
in commit 69c935baff. It is an XML/XMI file and therefore relies on
namespaces. The diagram contains references to the repositoryeditor
project to define the type of elements. When creating the diagram, the
repositoryeditor project was in the same workspace as the MediaStore
example and the *.aird file used "platform:/resource/" URIs. If you ran
a new Eclipse application with a new workspace and you imported only the
MediaStore, the old diagram could not be accessed. As a consequences, the
resources are now "platform:/plugin/" based, so you do not need to import
the repositoryeditor project for viewing the example diagram.